requirements.txt
python-dotenv==1.0.1
openai==1.59.3
streamlit==1.41.1
langchain==0.3.14
langchain-openai==0.2.14
langchain-community==0.3.14